pyproject.toml
[tool.poetry]
name = "firestone-lib"
version = "0.2.1"
description = "Library to help build OpenAPI, AsyncAPI and gRPC specs based off one or more resource json schema files"
authors = ["Erick Bourgeois <erick@jeb.ca>"]
license = "Apache 2.0"
readme = "README.md"
packages = [
    { include = "firestone_lib" }
]

[tool.poetry.dependencies]
python = ">=3.12,<=3.14"
click = ">=8.1.3"
Jinja2 = ">=3.1.2"
jsonschema = ">=4.17.3"
jsonref = ">=1.0.1"
pyyaml = ">=6.0"
pytest-cov = "^7.0.0"

[tool.poetry.group.build.dependencies]
black = ">=22.10.0"
pytest = ">=7.2.0"
pylint = ">=2.15.5"
pytest-mock = ">=3.10.0"
mock = ">=4.0.3"
pycodestyle = ">=2.9.1"
pytest-cov = ">=4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.pylint.'MESSAGES CONTROL']
max-line-length = 150
good-names = [
    "fh"
]
max-args = 8
disable = """
    logging-fstring-interpolation,
    fixme
"""
